Extracted out the classes, headers, and AmazonItem static methods
Moved the classes into their own respective files as well as the headers into a util then extracted out the scraping methods for the AmazonItem into static classes.
